Extract the remaining footer components
